[Gold V] Title: 합분해, Time: 0 ms, Memory: 2192 KB -BaekjoonHub

+++ 백준/Gold/2225. 합분해/합분해.cc
@@ -3,8 +3,8 @@ using namespace std;
 int N;
 int K;
 
-long long DP[210][210];
-long long mod = 1000000000;
+int DP[210][210];
+int mod = 1000000000;
 int main()
 {
 	ios::sync_with_stdio(false);
